requirements.txt
attrs==21.4.0
black==22.3.0
click==8.1.3
flake8==4.0.1
flake8-black==0.3.3
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
tomli==2.0.1
urllib3==1.26.9
